editoast: move `/version` in its own submodule
